refactor: services and models
