+++ src/models/index.ts
@@ -1,2 +1,3 @@
 export * from './settings'
 export * from './upload'
+export * from './users'
