--- src/services/auth/auth.models.ts
@@ -1,1 +1,0 @@
-export {}
