--- src/services/auth/auth.services.ts
@@ -1,1 +1,0 @@
-export const getGitHubUserData = () => {}
