+++ src/services/index.ts
@@ -1,4 +1,4 @@
-export * from './auth'
+export { default as AuthService } from './auth'
 export { default as SettingsService } from './settings'
 export { default as UploadService } from './upload'
-export * from './users'
+export { default as UsersService } from './users'
